requirements.txt
requests>=2.28.0
torch>=2.0.0
open-clip-torch>=2.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.8.0
faiss-cpu>=1.7.0
